+++ client/src/lib/currencyFormatter.js
@@ -1,0 +1,11 @@
+const currencyFormat = num => {
+  if (num) {
+    return num.toLocaleString("en-US", {
+      style: "currency",
+      currency: "USD"
+    });
+  }
+  return "N/A";
+};
+
+export default currencyFormat;
