+++ db/index.js
@@ -1,10 +1,7 @@
 const Pool = require("pg").Pool;
 
 const pool = new Pool({
-  user: process.env.PGUSER,
-  host: process.env.PGHOST,
-  database: process.env.PGDATABASE,
-  port: process.env.PGPORT
+  connectionString: process.env.DATABASE_URL
 });
 
 pool.on("error", (err, client) => {
